ODE physics - 1st pass - Correct null pointer referenes, concurrent data access, invalid logic

+++ WhiteCore/Physics/OpenDynamicsEngine/ODEDynamics.cs
@@ -1032,11 +1032,11 @@ namespace WhiteCore.Physics.OpenDynamicsEngine
 
             #region deflection
             //rotates body to direction of movement (linearMovement vector)
-			/* temporary disabled due to instabilities, needs to be rewritten
+			// temporary disabled due to instabilities, needs to be rewritten
             if(m_angularDeflectionTimescale < 300)
             {
                 float Dservo = 0.05f * m_angularDeflectionTimescale * m_angularDeflectionEfficiency;
-                float mag = (float)linearVelocity.LengthSquared();
+                float mag = linearVelocity.LengthSquared();
                 if(mag > 0.01f) {
                     linearVelocity.Y = -linearVelocity.Y;
                     linearVelocity *= rotq;
@@ -1052,7 +1052,7 @@ namespace WhiteCore.Physics.OpenDynamicsEngine
                     deflection *= Dservo;
                 }
             }
-			*/
+			//
             #endregion
 
             #region banking
